backend_python_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.3
httpx==0.27.0
pymongo==3.13.0
pyyaml==6.0.1
